pyproject.toml
[tool.poetry]
name = "cdk-ec2-cheap"
version = "0.1.0"
description = ""
authors = ["eisenhowerj <eisenhowerj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
constructs = "^10.0.39"
aws-cdk-lib = "^2.8.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
